dont cast to an intptr_t and then to an s32 dont cast an intptr_t to an s32 thats just silly

+++ src/render_player.c
@@ -1813,24 +1813,24 @@ void update_wheel_palette(Player* player, s8 playerId, s8 screenId, s8 arg3) {
 
             if (frameId <= 20) {
                 load_player_data_non_blocking(player,
-                                              (s32)(intptr_t) (gKartWheels0[player->characterId][groupId] +
+                                              (intptr_t) (gKartWheels0[player->characterId][groupId] +
                                                      (frameId * temp_num * 4) + ((temp_t2 >> 8) * 0x40)),
                                               D_802F1F80_WHEEL(arg3, screenId, playerId), 0x80);
             } else {
                 load_player_data_non_blocking(player,
-                                              (s32)(intptr_t) (gKartWheels1[player->characterId][groupId] +
+                                              (intptr_t) (gKartWheels1[player->characterId][groupId] +
                                                      (frameId - 21) * (temp_num * 4) + ((temp_t2 >> 8) * 0x40) + 0x600),
                                               D_802F1F80_WHEEL(arg3, screenId, playerId), 0x80);
             }
         } else {
             if (frameId == 0) {
                 load_player_data_non_blocking(player,
-                                              (s32)(intptr_t) (gKartWheels0[player->characterId][groupId] +
+                                              (intptr_t) (gKartWheels0[player->characterId][groupId] +
                                                      (frameId * temp_num * 4) + ((temp_t2 >> 8) * 0x40)),
                                               D_802F1F80_WHEEL(arg3, screenId, playerId), 0x80);
             } else {
                 load_player_data_non_blocking(player,
-                                              (s32)(intptr_t) (gKartWheels1[player->characterId][groupId] +
+                                              (intptr_t) (gKartWheels1[player->characterId][groupId] +
                                                      (frameId * temp_num * 4) + ((temp_t2 >> 8) * 0x40)),
                                               D_802F1F80_WHEEL(arg3, screenId, playerId), 0x80);
             }
